eld_backend/requirements.txt
annotated-types==0.7.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Django==5.1.4
django-cors-headers==4.9.0
djangorestframework==3.16.1
fastapi==0.115.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.0
linecache2==1.0.0
openai==1.57.0
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-multipart==0.0.19
reportlab==4.4.3
requests==2.32.3
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.41.3
tqdm==4.67.1
traceback2==1.4.0
typing_extensions==4.12.2
tzdata==2024.2
unittest2==1.1.0
urllib3==2.2.3
uvicorn==0.32.1
wheel==0.45.1
xmlrunner==1.7.7
